student_management/pyproject.toml
[project]
name = "student-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.2.1",
]
